Move "find parent job's config" fn to store
